# pi/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pi-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/shift-session/active": {
      "get": {
        "tags": [
          "shift"
        ],
        "summary": "Shift Session Active",
        "operationId": "shift_session_active_v1_shift_session_active_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "subject_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Type"
            }
          },
          {
            "name": "waiter_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Waiter Uuid"
            }
          },
          {
            "name": "cash_register_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cash Register Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftSessionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/shift-session/open": {
      "post": {
        "tags": [
          "shift"
        ],
        "summary": "Shift Session Open",
        "operationId": "shift_session_open_v1_shift_session_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftSessionOpenBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftSessionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/shift-session/{session_id}/close": {
      "post": {
        "tags": [
          "shift"
        ],
        "summary": "Shift Session Close",
        "operationId": "shift_session_close_v1_shift_session__session_id__close_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftSessionCloseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftSessionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/shift-session/{session_id}/receipt": {
      "post": {
        "tags": [
          "shift"
        ],
        "summary": "Shift Session Receipt",
        "operationId": "shift_session_receipt_v1_shift_session__session_id__receipt_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ShiftSessionReceiptBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftSessionEscposResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/shift-session/{session_id}/print": {
      "post": {
        "tags": [
          "shift"
        ],
        "summary": "Shift Session Print",
        "operationId": "shift_session_print_v1_shift_session__session_id__print_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftSessionCloseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftSessionPrintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShiftSessionCloseBody": {
        "properties": {
          "counted_cash_cents": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Counted Cash Cents"
          },
          "station_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Station Uuid"
          },
          "paper_width": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "80mm",
                  "58mm",
                  "53mm"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Width"
          },
          "charset": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pc858",
                  "pc850",
                  "cp1252",
                  "ascii"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Charset"
          }
        },
        "type": "object",
        "required": [
          "counted_cash_cents"
        ],
        "title": "ShiftSessionCloseBody"
      },
      "ShiftSessionEscposResponse": {
        "properties": {
          "cash_session_id": {
            "type": "integer",
            "title": "Cash Session Id"
          },
          "escpos_payload": {
            "type": "string",
            "title": "Escpos Payload"
          }
        },
        "type": "object",
        "required": [
          "cash_session_id",
          "escpos_payload"
        ],
        "title": "ShiftSessionEscposResponse"
      },
      "ShiftSessionOpenBody": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "subject_type": {
            "type": "string",
            "enum": [
              "waiter",
              "cash_register"
            ],
            "title": "Subject Type"
          },
          "waiter_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waiter Uuid"
          },
          "cash_register_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Register Uuid"
          },
          "operator_waiter_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Waiter Uuid"
          },
          "opening_balance_cents": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Opening Balance Cents",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "subject_type"
        ],
        "title": "ShiftSessionOpenBody"
      },
      "ShiftSessionPrintResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "print_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Print Job Id"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ShiftSessionPrintResponse"
      },
      "ShiftSessionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "subject_type": {
            "type": "string",
            "title": "Subject Type"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "opening_balance_cents": {
            "type": "integer",
            "title": "Opening Balance Cents"
          },
          "wallet_cents": {
            "type": "integer",
            "title": "Wallet Cents"
          },
          "total_cash_cents": {
            "type": "integer",
            "title": "Total Cash Cents"
          },
          "total_non_cash_cents": {
            "type": "integer",
            "title": "Total Non Cash Cents"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "event_id",
          "subject_type",
          "subject_name",
          "status",
          "opening_balance_cents",
          "wallet_cents",
          "total_cash_cents",
          "total_non_cash_cents"
        ],
        "title": "ShiftSessionRead"
      },
      "ShiftSessionReceiptBody": {
        "properties": {
          "counted_cash_cents": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Counted Cash Cents"
          },
          "paper_width": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "80mm",
                  "58mm",
                  "53mm"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Width"
          },
          "charset": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pc858",
                  "pc850",
                  "cp1252",
                  "ascii"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Charset"
          }
        },
        "type": "object",
        "title": "ShiftSessionReceiptBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}